Split TinySato/Printer.cs to TinySato/Printer/Behavior.cs
