+++ TinySato/TinySato.csproj
@@ -48,6 +48,7 @@
     <Compile Include="Communication\JobStatus.cs" />
     <Compile Include="ConnectionType.cs" />
     <Compile Include="Graphic.cs" />
+    <Compile Include="Printer\Behavior.cs" />
     <Compile Include="Printer\Density.cs" />
     <Compile Include="Printer\Sensor.cs" />
     <Compile Include="Printer.cs" />
